Add files via upload
In the app.py file, I designed the UI for the application using the Streamlit tool, where I integrated my speech recognition function and then ran it. 
In the speech.py file, I wrote the main logic of the code. In this file, I used the def recognize_speech() function for recognizing speech and also utilized cv2 and time to open a 3-second window and change it.